simplify baseapp and InitStater -> InitChainer
