--- types/genesis.go
@@ -1,6 +1,0 @@
-package types
-
-import "encoding/json"
-
-// function variable used to initialize application state at genesis
-type InitStater func(ctx Context, state json.RawMessage) Error
